+++ ChangeLog.markdown
@@ -4,6 +4,7 @@ Changelog for SqlDiff
 SqlDiff-0.0.3
 -------------
 * Fixed Issue #2: Missing default values
+* Added support for filtering using the --include and --exclude options
 
 SqlDiff-0.0.2
 -------------